requirements.txt
streamlit
openai
pymupdf
fpdf
pillow
openai>=0.27.0